# openapi_wasi.yaml
openapi: 3.0.3
info:
  title: API de Búsqueda de Inmuebles - WASI
  version: "2.0.0"
  description: |
    API REST para búsqueda avanzada de inmuebles usando datos sincronizados desde WASI.

servers:
  - url: http://localhost:5001
    description: Servidor local Flask (api_wasi.py)

paths:
  /:
    get:
      summary: Información de la API
      responses:
        '200':
          description: Información general de la API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiInfo'

  /estadisticas:
    get:
      summary: Estadísticas generales del dataset
      responses:
        '200':
          description: Estadísticas del dataset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadisticasGenerales'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /buscar:
    post:
      summary: Búsqueda avanzada de inmuebles
      description: |
        Búsqueda principal de inmuebles según múltiples criterios.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriteriosBusqueda'
      responses:
        '200':
          description: Resultados de la búsqueda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultadoBusqueda'
        '400':
          description: Criterios no proporcionados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /similares/{id}:
    get:
      summary: Inmuebles similares
      description: Devuelve inmuebles similares al inmueble de referencia.
      parameters:
        - name: id
          in: path
          required: true
          description: Índice interno del inmueble en el dataset (no el ID de WASI)
          schema:
            type: integer
        - name: n
          in: query
          required: false
          description: Número de inmuebles similares a retornar
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: Inmuebles similares
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilaresResponse'
        '400':
          description: ID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /inmueble/{id}:
    get:
      summary: Detalle de inmueble por ID WASI
      parameters:
        - name: id
          in: path
          required: true
          description: ID del inmueble en WASI
          schema:
            type: string
      responses:
        '200':
          description: Detalle de inmueble
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inmueble'
        '404':
          description: Inmueble no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tipos:
    get:
      summary: Tipos de inmuebles disponibles
      responses:
        '200':
          description: Lista de tipos de inmuebles y conteo por tipo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TiposResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /ciudades:
    get:
      summary: Ciudades disponibles
      responses:
        '200':
          description: Lista de ciudades, conteo y precio promedio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CiudadesResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /filtros-disponibles:
    get:
      summary: Filtros disponibles para búsqueda
      responses:
        '200':
          description: Filtros disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiltrosDisponibles'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /sincronizar:
    post:
      summary: Forzar sincronización con WASI
      responses:
        '200':
          description: Resultado de la sincronización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SincronizarResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ApiInfo:
      type: object
      properties:
        nombre:
          type: string
        version:
          type: string
        fuente_datos:
          type: string
        ultima_sincronizacion:
          type: string
          format: date-time
          nullable: true
        total_inmuebles:
          type: integer
        endpoints:
          type: object
          additionalProperties:
            type: string

    EstadisticasGenerales:
      type: object
      properties:
        total_inmuebles:
          type: integer
        ultima_sincronizacion:
          type: string
          format: date-time
          nullable: true
        precio_promedio:
          type: number
          format: float
        precio_minimo:
          type: number
          format: float
        precio_maximo:
          type: number
          format: float
        precio_mediana:
          type: number
          format: float
        distribucion_tipos:
          type: object
          additionalProperties:
            type: integer
        distribucion_ciudades:
          type: object
          additionalProperties:
            type: integer
        habitaciones_promedio:
          type: number
          format: float
        area_promedio:
          type: number
          format: float

    CriteriosBusqueda:
      type: object
      properties:
        tipo:
          type: string
          description: Tipo de inmueble (Casa, Apartamento, etc.)
        ciudad:
          type: string
        habitaciones_min:
          type: integer
        precio_max:
          type: number
          format: float
        tiene_piscina:
          type: boolean
        tipo_negocio:
          type: string
          description: venta, arriendo, etc.
        zona:
          type: string
        banos_min:
          type: integer
        area_min:
          type: number
      additionalProperties: true

    Inmueble:
      type: object
      properties:
        id:
          type: string
        codigo:
          type: string
        tipo:
          type: string
        tipo_negocio:
          type: string
        ciudad:
          type: string
        zona:
          type: string
        barrio:
          type: string
        direccion:
          type: string
        latitud:
          type: number
          format: float
        longitud:
          type: number
          format: float
        habitaciones:
          type: integer
        banos:
          type: integer
        garajes:
          type: integer
        area_construida:
          type: number
          format: float
        area_privada:
          type: number
          format: float
        area_total:
          type: number
          format: float
        estrato:
          type: integer
        antiguedad:
          type: integer
        piso:
          type: integer
        precio:
          type: number
          format: float
        precio_venta:
          type: number
          format: float
        precio_arriendo:
          type: number
          format: float
        administracion:
          type: number
          format: float
        estado:
          type: string
        disponible:
          type: boolean
        titulo:
          type: string
        descripcion:
          type: string
        url:
          type: string
          format: uri
        fecha_actualizacion:
          type: string
        tiene_piscina:
          type: boolean
        tiene_gimnasio:
          type: boolean
        tiene_parqueadero:
          type: boolean
        tiene_ascensor:
          type: boolean
        tiene_seguridad:
          type: boolean
      additionalProperties: true

    ResultadoBusqueda:
      type: object
      properties:
        total_encontrados:
          type: integer
        total_retornados:
          type: integer
        criterios:
          $ref: '#/components/schemas/CriteriosBusqueda'
        estadisticas:
          type: object
          properties:
            precio_promedio:
              type: number
              format: float
            precio_minimo:
              type: number
              format: float
            precio_maximo:
              type: number
              format: float
        resultados:
          type: array
          items:
            $ref: '#/components/schemas/Inmueble'
        mensaje:
          type: string

    SimilaresResponse:
      type: object
      properties:
        inmueble_referencia:
          $ref: '#/components/schemas/Inmueble'
        similares_encontrados:
          type: integer
        similares:
          type: array
          items:
            $ref: '#/components/schemas/Inmueble'

    TiposResponse:
      type: object
      properties:
        tipos:
          type: array
          items:
            type: string
        conteo:
          type: object
          additionalProperties:
            type: integer

    CiudadesResponse:
      type: object
      properties:
        ciudades:
          type: array
          items:
            type: string
        conteo:
          type: object
          additionalProperties:
            type: integer
        precio_promedio:
          type: object
          additionalProperties:
            type: number
            format: float

    FiltrosDisponibles:
      type: object
      properties:
        tipos:
          type: array
          items:
            type: string
        ciudades:
          type: array
          items:
            type: string
        zonas:
          type: array
          items:
            type: string
        tipo_negocio:
          type: array
          items:
            type: string
        habitaciones:
          type: array
          items:
            type: integer
        banos:
          type: array
          items:
            type: integer
        caracteristicas_booleanas:
          type: array
          items:
            type: string
        rangos_numericos:
          type: object
          properties:
            precio:
              type: object
              properties:
                min:
                  type: number
                  format: float
                max:
                  type: number
                  format: float
            area_total:
              type: object
              properties:
                min:
                  type: number
                  format: float
                max:
                  type: number
                  format: float
            area_construida:
              type: object
              properties:
                min:
                  type: number
                  format: float
                max:
                  type: number
                  format: float

    SincronizarResponse:
      type: object
      properties:
        mensaje:
          type: string
        timestamp:
          type: string
          format: date-time
        total_inmuebles:
          type: integer

    ErrorResponse:
      type: object
      properties:
        error:
          type: string